+++ drydock/drydock.py
@@ -4,12 +4,13 @@ __doc__ = """DryDock v{}
 
 Usage:
     drydock construct [-s] <specification>
+    drydock deconstruct [-s] <specification>
     drydock prepare
     drydock master [options] <name>
     drydock --help | --version
 
 Options:
-    -s --supervisor         Write a supervisor config.
+    -s --supervisor         Include the supervisor config.
 
     -p --http=<http>        HTTP Port. [default: 80]
     -l --https=<https>      HTTPS Port. [default: 443]
@@ -19,9 +20,10 @@ Options:
     -v --version            Show current version.
 
 Commands:
-    construct   Construct the given specification.
-    prepare     Prepare docker for constructing a drydock specification.
-    master      Create a container that can run a drydock specification.
+    construct    Construct the given specification.
+    deconstruct  Deconstruct the given specification.
+    prepare      Prepare docker for constructing a drydock specification.
+    master       Create a container that can run a drydock specification.
 
 DryDock takes a simple (YAML) specification file then can construct and
 configure a cluster of docker containers. DryDock will automatically setup a
@@ -43,6 +45,12 @@ def main():
     if args["construct"]:
         with open(args["<specification>"], 'r') as drydock:
             construction.construct(
+                MetaContainer(**yaml.load(drydock.read())),
+                args["--supervisor"])
+
+    elif args["deconstruct"]:
+        with open(args["<specification>"], 'r') as drydock:
+            construction.deconstruct(
                 MetaContainer(**yaml.load(drydock.read())),
                 args["--supervisor"])