+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 ++++++++++++++++++
 
 * Added: supervisor config writing is now an option.
+* Added: deconstruct command to remove a specification.
 
 0.1.0 (25-02-2014)
 ++++++++++++++++++